Time_Series_Price_Prediction/requirements.txt
numpy==2.1.4
pandas==2.2.2
yfinance==0.2.40
scikit-learn==1.6.1
PyYAML==6.0.2
